// files/src/react/modules/stations/stations.enum.ts
export enum VendorStationStatus {
  active = 'active',
  inactive = 'inactive',
}

export enum VendorTypes {
  sapura = 'sapura',
  sentinel = 'sentinel',
  setel = 'setel',
}

export enum StationStatus {
  COMIN_SOON = 'coming-soon',
  MAINTENANCE = 'maintenance',
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum PumpStatus {
  INACTIVE = 'inactive',
  ACTIVE = 'active',
  MAINTENANCE = 'maintenance',
  IMPORTED = 'imported',
}